Fix memory leak in deletetexteditortext()
Opening the global text editor (by pressing the ! key) allocates a new
stringvector to hold the text, and closing the text editor frees all the
nodes within that stringvector via deletetexteditortext(). However, the
stringvector struct *itself* is never properly freed. This commit
updates deletetexteditortext() to free it.

This commit also updates editprogram() to allocate its stringvector on
the heap instead of on the stack, so that calling free() on an editor's
stringvector is always a valid thing to do.